Fix ColorLerp casting float to byte which caused black transitions
The Color struct uses float values (0.0-1.0), but ColorLerp was
incorrectly casting the lerp results to byte before passing them
to the Color constructor. This caused values like 0.95 to become 0,
resulting in black colors during transitions.

+++ src/Sandbox.UI/Styles/BaseStyles.Utility.cs
@@ -27,10 +27,10 @@ namespace Sandbox.UI
 		{
 			t = Math.Clamp(t, 0f, 1f);
 			return new Color(
-				(byte)MathX.Lerp(a.r, b.r, t),
-				(byte)MathX.Lerp(a.g, b.g, t),
-				(byte)MathX.Lerp(a.b, b.b, t),
-				(byte)MathX.Lerp(a.a, b.a, t)
+				MathX.Lerp(a.r, b.r, t),
+				MathX.Lerp(a.g, b.g, t),
+				MathX.Lerp(a.b, b.b, t),
+				MathX.Lerp(a.a, b.a, t)
 			);
 		}
